backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-cors==4.0.1
psycopg[binary]
python-dotenv==1.0.1
werkzeug==3.0.3
